// HealthCheck.Frontend/src/generated/Models/Core/HCJobsGetPagedHistoryRequestModel.ts
//     This code was generated by a Reinforced.Typings tool. 
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface HCJobsGetPagedHistoryRequestModel
{
	SourceId: string;
	JobId: string;
	PageIndex: number;
	PageSize: number;
}