// HealthCheck.Frontend/src/generated/Models/Core/HCJobDefinitionWithSourceViewModel.ts
//     This code was generated by a Reinforced.Typings tool. 
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import { HCJobDefinitionViewModel } from './HCJobDefinitionViewModel';

export interface HCJobDefinitionWithSourceViewModel
{
	SourceId: string;
	Definition: HCJobDefinitionViewModel;
}